fix(circle): fix circle local count updater
